% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kegg_up_down_link.R
\name{add_kegg_up_down_link}
\alias{add_kegg_up_down_link}
\title{add links for the KEGG enrichment table.}
\usage{
add_kegg_up_down_link(keggEnrichTable, logFC, diff_gene_list,
  studyID = "test")
}
\arguments{
\item{keggEnrichTable}{a data.frame which must have PathwayID and pathway_name columns}

\item{logFC}{A numeric vector, the names for the vector are gene lists.}

\item{diff_gene_list}{a vector which contain the significantly DEGs. symbol or entrez ID.}

\item{studyID}{A standard study ID,default:test}
}
\value{
write a html file with links for the keggEnrichTable
}
\description{
KEGG enrichment table comes from the stantard hyperG test. the PathwayID and pathway_name columns  are needed.
 logFC <- DEG$logFC;names(logFC) <- DEG$symbol
 diff_gene_list also comes from DEG results.
}
\examples{
#' add_kegg_up_down_link(keggEnrichTable , DEG , diff_gene_list,studyID)
}
\keyword{keggAnno}

