% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_enrichment.R
\name{batch_enrichment}
\alias{batch_enrichment}
\title{enrichment for GO/KEGG}
\usage{
batch_enrichment(diff_gene_file, all_genes_file, prefix = "test",
  GOstats = F)
}
\arguments{
\item{diff_gene_file}{just one column, gene HUGO symbol or entrez ID , no need for column name.}

\item{all_genes_file}{the same to the diff_gene_file, contain more genes than it.}

\item{prefix}{The prefix for the output files.}

\item{GOstats}{whether use GOstats or not , defaults:F}
}
\value{
write enrichment results into files.
}
\description{
enrichment for GO/KEGG
}
\examples{
#'  batch_enrichment('diff_gene_file.txt','all_genes_file.txt')


}
\keyword{batch_enrichment}

