% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_t_test.R
\name{batch_t_test}
\alias{batch_t_test}
\title{do batch t.test for a matrix}
\usage{
batch_t_test(dat = matrix(rnorm(300), 30, 10), group1 = 1:5,
  group2 = 6:10)
}
\arguments{
\item{dat}{An expression matrix which rownames are genes and columns are samples}

\item{group1}{defined which columns belong to the first group as control}

\item{group2}{defined which columns belong to the second group}
}
\value{
a data.frame just like topTable
}
\description{
do batch t.test for a matrix
}
\examples{
#'  batch_t_test(exprSet ,4:7,1:3)


}
\keyword{batch_enrichment}

