% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_plot.R
\name{cc_plot}
\alias{cc_plot}
\title{plot overall/within group mean correlation coefficient of samples}
\usage{
cc_plot(expr, meta, ps = 8, fs = 23)
}
\arguments{
\item{expr}{a expression matrix, samples in columns, expression values in rows:}

\item{meta}{a vector, for group, sample and color of group}
}
\value{
correlation matrix and correlation figure
}
\description{
A function to calculate overall/within group mean correlation coefficient of samples for the Basic QC
}
\examples{
#' cc_plot()
}
\keyword{correlation}

