% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshData.R
\name{create_dup_exprSet}
\alias{create_dup_exprSet}
\title{create the duplicated expression matrix}
\usage{
create_dup_exprSet(refresh = F)
}
\arguments{
\item{refresh}{Don't use this function if not neccessary}
}
\description{
the expression matrix come from the data 'sCLLex' in R package 'CLL'
duplicated because one gene symbol corresponding to multiple probeset ID in hgu95av2.db
dup_exprSet is a data.frame,  the first column is gene symbol, which needed to remove duplicated.
}
\keyword{refresh}

