% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshData.R
\name{create_example_exprSet}
\alias{create_example_exprSet}
\title{create expression matrix,rowname are unique genes, colname are samples}
\usage{
create_example_exprSet(refresh = F)
}
\arguments{
\item{refresh}{Don't use this function if not neccessary}
}
\description{
example_exprSet comes from rmDupID(dup_exprSet),also the data 'sCLLex' in R package 'CLL'
}
\keyword{refresh}

