% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshData.R
\name{create_genesets_list}
\alias{create_genesets_list}
\title{create a list of gene sets for  enzyme/protein complex/ development/immune}
\usage{
create_genesets_list(refresh = F)
}
\arguments{
\item{refresh}{Don't use this function if not neccessary}
}
\description{
PTM: Post-translational modification, such methyl-,  acethyl-,  phospha-
SWI/SNF Chromatin Remodeling Complexes and PRC family members
development associated genes
Immune-Related Genes
transcription factor genesets
other genesets
}
\keyword{refresh}

