% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGSEAinput.R
\name{createGSEAinput}
\alias{createGSEAinput}
\title{create gct file and cls file}
\usage{
createGSEAinput(studyID = "GSE1009", exprSet = example_exprSet, group_list,
  destdir = ".")
}
\arguments{
\item{studyID}{A standard study ID in GEO: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE49822}

\item{exprSet}{a expression matrix, which columns are sample,rows are HUGO gene symbols.}

\item{group_list}{a vector,as long as the col number for the expression matrix,which describe the group for the samples in exprSet}

\item{destdir}{where to store the files just download.}
}
\value{
A expresstionSet,S4 object in R, which contain: assayData,phenoData,featureData
}
\description{
create gct file and cls file for GSEA according the expression matrix and group information
}
\examples{
#' createGSEAinput('GSE1009')
}
\keyword{downGSE}

