% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_DEG_2groups.R
\name{do_DEG_2groups}
\alias{do_DEG_2groups}
\title{Do differential expression analysis for the expression matrix}
\usage{
do_DEG_2groups(prefix = "GSE1009", exprSet = example_exprSet, group_list,
  method, destdir = ".")
}
\arguments{
\item{prefix}{The prefix for all of the output files.( we don't need it now actually)}

\item{exprSet}{Matrix with microarray expression values.}

\item{group_list}{Factors for two groups that are tested for differential expression.}

\item{method}{which method do you choose, limma,samr,rowttest in genefilter,resamplingPvalues in BioNet,t.test}

\item{destdir}{where to store the files just download.( we don't need it now actually)}
}
\value{
a data.frame just like topTable
}
\description{
please make sure that there are just two group for you expression matrix, you should modify the code in this function if there's more than 2 group.
I highly suggest you to use Limma to analysis
}
\examples{
#' do_DEG_2groups('CLL',exprSet,group_list,method='limma')
}
\keyword{do_DEG_2groups}

