% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downGSE.R
\name{downGSE}
\alias{downGSE}
\title{download the data from GEO database}
\usage{
downGSE(studyID = "GSE1009", destdir = ".")
}
\arguments{
\item{studyID}{A standard study ID in GEO: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE1009}

\item{destdir}{where to store the files just download.}
}
\value{
A expresstionSet,S4 object in R, which contain: assayData,phenoData,featureData
}
\description{
Given a GSE study ID, just like GSE1009, this function will download the eSet object and write the expression matrix and phenotype information.
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE1009
}
\examples{
#' downGSE('GSE1009')
}
\keyword{downGSE}

