% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_boxplot_gene.R
\name{draw_boxplot_gene}
\alias{draw_boxplot_gene}
\title{draw boxplot for a specific gene's values between two groups}
\usage{
draw_boxplot_gene(gene = "VCX3A", exprSet, group_list)
}
\arguments{
\item{gene}{specific a gene which should be the rownames of exprSet}

\item{exprSet}{An expression matrix which rownames are genes and columns are samples}

\item{group_list}{defined which sample belong to which group}
}
\value{
create a boxplot
}
\description{
draw boxplot for a specific gene's values between two groups
}
\examples{
#'  draw_boxplot_gene(x,exprSet,group_list)


}
\keyword{batch_enrichment}

