% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_DEG.R
\name{format_DEG}
\alias{format_DEG}
\title{write some files from the DEG results from topTable(limma)}
\usage{
format_DEG(DEG, prefix = "test", logFC_cutoff = 0, pvalue_cutoff = 0,
  padjust_cutoff = 0, GOstats = F)
}
\arguments{
\item{DEG}{DEG=topTable(fit,coef=2,adjust='BH')}

\item{prefix}{The prefix for all of the output files.}

\item{logFC_cutoff}{The cutoff for logFC    to choose DEG, default:mean(abs(DEG$logFC)) + 2*sd(abs(DEG$logFC))}

\item{pvalue_cutoff}{The cutoff for pvalue   to choose DEG, default:0.05}

\item{padjust_cutoff}{The cutoff for p.adjust to choose DEG, default:none}
}
\value{
write some files
}
\description{
write some files from the DEG results from topTable(limma)
}
\examples{
#'  format_DEG(DEG)


}
\keyword{format_DEG}

