% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GSE_links.R
\name{get_GSE_links}
\alias{get_GSE_links}
\title{Get the links for GSE studyID}
\usage{
get_GSE_links(studyID = "GSE1009", down = F, destdir = "./")
}
\arguments{
\item{studyID}{A standard study ID}

\item{destdir}{where to store the files download according to current GSE studyID}
}
\value{
write some files
}
\description{
Get the links for GSE studyID
}
\examples{
#' get_GSE_links();get_GSE_links(studyID='GSE1009',down=T);get_GSE_links('GSE42872')


}
\keyword{get_GSE_links}

