% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{Main GSEA Analysis Function that implements the entire methodology}
\usage{
GSEA(input.ds, input.cls, gene.ann = "", gs.db, gs.ann = "",
  output.directory = "", doc.string = "GSEA.analysis",
  non.interactive.run = F, reshuffling.type = "sample.labels",
  nperm = 1000, weighted.score.type = 1, nom.p.val.threshold = -1,
  fwer.p.val.threshold = -1, fdr.q.val.threshold = 0.25, topgs = 10,
  adjust.FDR.q.val = F, gs.size.threshold.min = 25,
  gs.size.threshold.max = 500, reverse.sign = F, preproc.type = 0,
  random.seed = 123456, perm.type = 0, fraction = 1, replace = F,
  save.intermediate.results = F, OLD.GSEA = F,
  use.fast.enrichment.routine = T)
}
\arguments{
\item{input.ds:}{Input gene expression Affymetrix dataset file in RES or GCT format}

\item{input.cls:}{Input class vector (phenotype) file in CLS format}

\item{gene.ann.file:}{Gene microarray annotation file (Affymetrix Netaffyx *.csv format) (default: none)}

\item{gs.file:}{Gene set database in GMT format}

\item{output.directory:}{Directory where to store output and results (default: .)}

\item{doc.string:}{Documentation string used as a prefix to name result files (default: "GSEA.analysis")}

\item{non.interactive.run:}{Run in interactive (i.e. R GUI) or batch (R command line) mode (default: F)}

\item{reshuffling.type:}{Type of permutation reshuffling: "sample.labels" or "gene.labels" (default: "sample.labels")}

\item{nperm:}{Number of random permutations (default: 1000)}

\item{weighted.score.type:}{Enrichment correlation-based weighting: 0=no weight (KS), 1=standard weigth, 2 = over-weigth (default: 1)}

\item{nom.p.val.threshold:}{Significance threshold for nominal p-vals for gene sets (default: -1, no thres)}

\item{fwer.p.val.threshold:}{Significance threshold for FWER p-vals for gene sets (default: -1, no thres)}

\item{fdr.q.val.threshold:}{Significance threshold for FDR q-vals for gene sets (default: 0.25)}

\item{topgs:}{Besides those passing test, number of top scoring gene sets used for detailed reports (default: 10)}

\item{adjust.FDR.q.val:}{Adjust the FDR q-vals (default: F)}

\item{gs.size.threshold.min:}{Minimum size (in genes) for database gene sets to be considered (default: 25)}

\item{gs.size.threshold.max:}{Maximum size (in genes) for database gene sets to be considered (default: 500)}

\item{reverse.sign:}{Reverse direction of gene list (pos. enrichment becomes negative, etc.) (default: F)}

\item{preproc.type:}{Preprocessing normalization: 0=none, 1=col(z-score)., 2=col(rank) and row(z-score)., 3=col(rank). (default: 0)}

\item{random.seed:}{Random number generator seed. (default: 123456)}

\item{perm.type:}{Permutation type: 0 = unbalanced, 1 = balanced. For experts only (default: 0)}

\item{fraction:}{Subsampling fraction. Set to 1.0 (no resampling). For experts only (default: 1.0)}

\item{replace:}{Resampling mode (replacement or not replacement). For experts only (default: F)}

\item{OLD.GSEA:}{if TRUE compute the OLD GSEA of Mootha et al 2003}

\item{use.fast.enrichment.routine:}{if true it uses a faster version to compute random perm. enrichment "GSEA.EnrichmentScore2"}
}
\value{
produce some figures
}
\description{
This is a methodology for the analysis of global molecular profiles called Gene Set Enrichment Analysis (GSEA). It determines
 whether an a priori defined set of genes shows statistically significant, concordant differences between two biological
 states (e.g. phenotypes). GSEA operates on all genes from an experiment, rank ordered by the signal to noise ratio and
 determines whether members of an a priori defined gene set are nonrandomly distributed towards the top or bottom of the
 list and thus may correspond to an important biological process. To assess significance the program uses an empirical
 permutation procedure to test deviation from random that preserves correlations between genes.
}
\examples{
#'  GSEA(input.ds =  input.ds.file,input.cls =input.cls.file ,gs.db = gs.db.file ,output.directory='./', reshuffling.type      = "gene.labels")



}
\keyword{GSEA}

