% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperGtest_jimmy.R
\name{hyperGtest_jimmy}
\alias{hyperGtest_jimmy}
\title{Hypergeometric Tests for GO/KEGG test by jimmy}
\usage{
hyperGtest_jimmy(GeneID2Path = GeneID2kegg_list,
  Path2GeneID = kegg2GeneID_list,
  diff_gene = sample(unique(hgu95av2_id$gene_id), 500),
  universeGeneIds = unique(hgu95av2_id$gene_id))
}
\arguments{
\item{GeneID2Path}{a list which one entrez gene id to multiple pathway id,default:GeneID2kegg_list}

\item{Path2GeneID}{a list which one pathway id to multiple entrezgene id,default:kegg2GeneID_list}

\item{diff_gene}{a vector which contain the significantly DEG list, abouth 500~1000 genes(entrez gene id ),default:sample(unique(hgu95av2_id$gene_id),500)}

\item{universeGeneIds}{a vector which contain the backgroud gene list,probably 20,000 genes , default:unique(hgu95av2_id$gene_id)}
}
\value{
a data.frame, each row is a Hypergeometric Tests result for each pathway .
}
\description{
Given GeneID2Path,Path2GeneID,diff_gene,universeGeneIds, this function will compute  Hypergeometric Tests for each path (GO/KEGG),
It can't hold on the structure of the GO graph, just a simple path .
}
\examples{
#' hyperGtest_jimmy(),hyperGtest_jimmy()
}
\keyword{hyperGtest}

