% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggAnno.R
\name{keggAnno}
\alias{keggAnno}
\title{annotate KEGG pathway for entrez id or symbol}
\usage{
keggAnno(geneLists = c(1, 2, 9), fileType = "html", fileName = "gene",
  multiple = T)
}
\arguments{
\item{geneLists}{a character vector for gene entrez ID list,default: c(1,2,9)}

\item{fileType}{choose whether annotate gene name     for the gene id or not , default : T}

\item{fileName}{choose whether annotate gene map      for the gene id or not , default : F}

\item{multiple}{choose whether annotate gene ensembl  for the gene id or not , default : F}
}
\value{
write a html or csv files for the annotation results.
}
\description{
somethins
}
\examples{
#' keggAnno('TP53',fileName='TP53',multiple = F),keggAnno('TP53')
}
\keyword{keggAnno}

