% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_heatmap.R
\name{pathway_heatmap}
\alias{pathway_heatmap}
\title{draw heatmap for a single pathway or a batch of pahways}
\usage{
pathway_heatmap(exprSet, pathwayID = "00910", search = "NA", all = F)
}
\arguments{
\item{exprSet}{a expression matrix, which columns are sample,rows are HUGO gene symbols.}

\item{pathwayID}{a character just like 00910(means 'Nitrogen metabolism'),default:00910}

\item{search}{Generate heatmap for the kegg pathway which related the character you search.}

\item{all}{generate heatmap for all kegg pathway if set this parameter as TURE. default:F}
}
\value{
generate QC figures for  expression matrix
}
\description{
you can point the pathway you are interested in or a list of related pathway ,such immune or sinal pathway
}
\examples{
pathway_heatmap(example_exprSet)
}
\keyword{QC}

