% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ES_score.R
\name{plot_ES_score}
\alias{plot_ES_score}
\title{re-draw ES-score figures for GSEA}
\usage{
plot_ES_score(Ng = 12, N = 34688, phen1 = "control", phen2 = "case",
  Obs.RES, Obs.indicator, obs.s2n, size.G, gs.names, Obs.ES, Obs.arg.ES,
  Obs.ES.index)
}
\arguments{
\item{Ng}{The number of the gene　sets, (eg: 200~300 kegg pathway)}

\item{N}{The number of the genes (probably 20,000~30,000 genes)}

\item{phen1}{The name for the first phenotype (default:control)}

\item{phen2}{The name for the second phenotype (default:case)}

\item{Obs.RES}{A matrix, running ES score for each gene in specific gene Set.}

\item{Obs.indicator}{A matrix, 0/1 shows whether a gene belong to a gene Set.}

\item{obs.s2n}{A vector of sigal to noise value for each gene (sorted, and Z-score)}

\item{size.G}{A vector for the size of each geneSet}

\item{gs.names}{A vector for the name of each geneSet}

\item{Obs.ES}{A vector for the maximal ES score of each gene Set.}

\item{Obs.arg.ES}{A vector for the order of the gene shows maximal ES score of each gene Set.}
}
\value{
write some files
}
\description{
It's very complicate,please don't use this unless you have a clear understanding of the mechanism for the GSEA
especially for the ES core ,the running ES core, the gene sets !
you need to prepare many data by yourself, I really can't help you with them.
#'
}
\examples{
#'  plot_ES_score(Ng=12,N=34688,phen1='control',phen2='case',Obs.RES,Obs.indicator,obs.s2n,size.G,gs.names,Obs.ES,Obs.arg.ES,Obs.ES.index)


}
\keyword{plot_ES_score}

