% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probesetAnno.R
\name{probesetAnno}
\alias{probesetAnno}
\title{annotate entrez gene id for the given probeset}
\usage{
probesetAnno(probeList = c("1000_at", "1001_at"), platform = "hgu95av2")
}
\arguments{
\item{probeList}{a character vector for probeset ID list,default:c('1000_at','1001_at')}

\item{platform}{the platform for probeList, So far, we just accept hgu95av2/hgu133a/hgu133b/hgu133plus2}
}
\value{
a data.frame which has two column, first is probeset ID , second is entrez gene id .
}
\description{
somethins
}
\keyword{microarray}

