% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCexpressionMatrix.R
\name{QCexpressionMatrix}
\alias{QCexpressionMatrix}
\title{Do basic QC for a expression matrix.}
\usage{
QCexpressionMatrix(exprSet, group_list = "NO", project = "test")
}
\arguments{
\item{exprSet}{a expression matrix, which columns are sample,rows are HUGO gene symbols.}

\item{group_list}{a vector,as long as the col number for the expression matrix,which describe the group for the samples in exprSet}
}
\value{
generate QC figures for  expression matrix
}
\description{
Basic QC include:boxplot and hist figures for expression value distribution
                 correlations among the samples by clustering
}
\examples{
#' QCexpressionMatrix(example_exprSet,group_list,project='test')
}
\keyword{QC}

