% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshData.R
\name{refreshMicroarrayData}
\alias{refreshMicroarrayData}
\title{extract probeset ID and entrez gene id from R packages}
\usage{
refreshMicroarrayData(refresh = F)
}
\arguments{
\item{refresh}{Don't use this function if not neccessary}
}
\description{
So far, we just extract data from R packages:hgu95av2/hgu133a/hgu133b/hgu133plus2
}
\keyword{refresh}

