% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshData.R
\name{refreshOrgHsData}
\alias{refreshOrgHsData}
\title{extract gene IDs from org.Hs.eg.db}
\usage{
refreshOrgHsData(refresh = F)
}
\arguments{
\item{refresh}{Don't use this function if not neccessary}
}
\description{
So far, we just extract gene entrez ID,ensembl ID,HUGO symbol,and locus.
}
\keyword{refresh}

