% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDupID.R
\name{rmDupID}
\alias{rmDupID}
\title{remove duplicat gene entrez ID or symbol for a expression matrix.}
\usage{
rmDupID(dup_exprSet)
}
\arguments{
\item{dup_exprSet}{a expression matrix or data.frame which the  first column is the ID needed to remove duplicate.}
}
\value{
just  expression matrix which colname is the unique ID, and value is numeric.
}
\description{
sometimes, we will get a expression matrix, which one gene has many values in each sample, because we design many probeset for this  gene
most of time, we just need one value for each gene, so we should remove the duplicate value by choose the biggest.
the duplicted gene can be entrez gene ID or HUGO gene symbol
}
\examples{
#' exprSet <- rmDupID(dup_exprSet)

}
\keyword{rmDupID}

