% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreshData.R
\name{update_kegg}
\alias{update_kegg}
\title{update kegg data}
\usage{
update_kegg(refresh = F)
}
\arguments{
\item{refresh}{Don't use this function if not neccessary}
}
\description{
It's not a automately function, please don't run it. it will read the files ind data folder, which is 'kegg2geneID.txt' and 'kegg_hierarchical.txt'
you must make sure the files exists in the data/KEGG_update folder.
It will process the files and store the GeneID2kegg_list,kegg2GeneID_list,keggID2geneID_df,kegg2name to the Rdata.
GeneID2kegg_list and kegg2GeneID_list are list which one key has multiple values.
keggID2geneID_df is a data.frame , 2 columns : 'gene_id','path_id'
kegg2name is a data.frame, which has 4 columns, which is :parent1 parent2 pathway_id pathway_name
}
\keyword{refresh}

