% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcanic_DEG.R
\name{Volcanic_DEG}
\alias{Volcanic_DEG}
\title{write some files from the DEG results from topTable(limma)}
\usage{
Volcanic_DEG(DEG, imageType = "png", prefix = "test")
}
\arguments{
\item{DEG}{DEG=topTable(fit,coef=2,adjust='BH')}

\item{imageType}{choose png or pdf or EMF}

\item{prefix}{A standard study ID}
}
\value{
plot 3 kind of volcanic figures
}
\description{
write some files from the DEG results from topTable(limma)
}
\examples{
#'  Volcanic_DEG(DEG)


}
\keyword{Volcanic_DEG}

