% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_pheatmap.R
\name{batch_pheatmap}
\alias{batch_pheatmap}
\title{Do batch heatmap for interested gene sets.}
\usage{
batch_pheatmap(exprSet, needGeneList)
}
\arguments{
\item{exprSet}{a expression matrix, which columns are sample,rows are HUGO gene symbols.}

\item{needGeneList}{a list of interested gene sets, each gene set is a vector of genes, the name will be used by out figures.}
}
\value{
generate heatmap figures for each gene sets.
}
\description{
extract the corresponding expression matrix for each gene set , and draw heatmap for the gene set expression matrix.
The name of the gene set will be used for the name of the heatmap figure
the rowname must be HUGO gene symbols
}
\examples{
#' batch_pheatmap(),batch_pheatmap()
}
\keyword{heatmap}

