% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_t_test.R
\name{batch_t_test}
\alias{batch_t_test}
\title{do batch t.test for a matrix}
\usage{
batch_t_test(dat = matrix(rnorm(300), 30, 10), group_list = factor(c(rep(1,
  5), rep(2, 5))))
}
\arguments{
\item{dat}{Matrix with microarray expression values.}

\item{group_list}{Factors for two groups that are tested for differential expression.}
}
\value{
a data.frame just like topTable
}
\description{
do batch t.test for a matrix
}
\examples{
#'  batch_t_test()


}
\keyword{batch_t_test}

