% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_DEG.R
\name{format_DEG}
\alias{format_DEG}
\title{write some files from the DEG results from topTable(limma)}
\usage{
format_DEG(DEG, prefix = "test", GOstats = F)
}
\arguments{
\item{DEG}{DEG=topTable(fit,coef=2,adjust='BH')}

\item{prefix}{The prefix for all of the output files.}
}
\value{
write some files
}
\description{
write some files from the DEG results from topTable(limma)
}
\examples{
#'  format_DEG(DEG)


}
\keyword{format_DEG}

