% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnno.R
\name{geneAnno}
\alias{geneAnno}
\title{annotate gene name or map for entrez id or symbol}
\usage{
geneAnno(geneLists = c(1, 2, 9), name = T, map = F, ensembl = F,
  accnum = F)
}
\arguments{
\item{geneLists}{a character vector for gene entrez ID list,default: c(1,2,9)}

\item{name}{choose whether annotate gene name     for the gene id or not , default : T}

\item{map}{choose whether annotate gene map      for the gene id or not , default : F}

\item{ensembl}{choose whether annotate gene ensembl  for the gene id or not , default : F}

\item{accnum}{choose whether annotate gene accnum   for the gene id or not , default : F}
}
\value{
a data.frame which has more than 2 column
}
\description{
somethins
}
\keyword{geneAnno}

