% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbol_exprSet.R
\name{get_symbol_exprSet}
\alias{get_symbol_exprSet}
\title{extract the expression matrix from a eSet object}
\usage{
get_symbol_exprSet(eSet, platformDB = "hgu133plus2.db")
}
\arguments{
\item{eSet}{A expresstionSet,S4 object in R, which contain: assayData,phenoData,featureData}

\item{platformDB}{which microarray platform does the eSet use ?}
}
\value{
expression matrix, rowname is the gene HUGO symbol
}
\description{
not only extract the expression matrix, but change the probeset ID to gene HUGO symbol ,and remove the duplicated genes .
}
\examples{
#' exprSet <- get_symbol_exprSet(K27M_WT_eSet)


}
\keyword{get_symbol_exprSet}

